make module contents service
