+++ packages/host/app/services/card-type-service.ts
@@ -62,8 +62,19 @@ export default class CardTypeService extends Service {
 
   private typeCache: Map<string, Type> = new Map();
   private moduleInfoCache: Map<string, ModuleInfo> = new Map();
+  private loaderNonce: object | undefined;
+
+  invalidateAllCaches(): void {
+    this.typeCache.clear();
+    this.moduleInfoCache.clear();
+  }
 
   async assembleType(definition: typeof BaseDef): Promise<Type> {
+    // This should go away when we move to an architecture where NO loader reset is required
+    if (this.loaderNonce !== this.loaderService.loader) {
+      this.invalidateAllCaches();
+      this.loaderNonce = this.loaderService.loader;
+    }
     let maybeType = await this.toType(definition, this.loaderService.loader);
     if (isCodeRefType(maybeType)) {
       throw new Error(`bug: should never get here`);
